requirements.txt
# Core AI Framework
crewai[tools]>=0.177.0,<1.0.0

# Web Frameworks
flask>=2.0.0,<3.0.0
streamlit>=1.35.0
streamlit-authenticator>=0.2.4
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.3
streamlit-plotly-events>=0.2.0

# Data Visualization
plotly>=5.17.0
altair>=5.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0

# PDF and Document Processing
reportlab>=4.0.0
fpdf2>=2.7.0
streamlit-pdf-viewer>=1.0.0
streamlit-drawable-canvas>=0.9.3

# Network Analysis and Visualization
networkx>=3.1
pyvis>=0.3.2
streamlit-agraph>=0.0.45

# Database and Storage
sqlalchemy>=2.0.0
pymongo>=4.5.0
redis>=4.6.0

# Authentication and Security
pyjwt>=2.8.0
bcrypt>=4.0.0
passlib>=1.7.4
cryptography>=41.0.0

# Business Intelligence
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Cloud and Infrastructure
boto3>=1.28.0
kubernetes>=27.0.0
terraform-compliance>=1.3.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
jsonschema>=4.19.0
toml>=0.10.0
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.5.0
loguru>=0.7.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0